requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
PyQt6>=6.4.0
mediapipe>=0.10.0
Pillow>=10.0.0
pycups>=2.0.1 